requirements.txt
fpdf==1.7.2
arabic_reshaper
python-bidi
